Move shortest_path to its own module
